[WIP] First pass at refactoring the shared network step to remove vSphere-specific items into a subclass
