+++ pkg/v1/tkg/web/src/app/shared/service/Messenger.ts
@@ -7,12 +7,12 @@ import { Injectable } from '@angular/core';
 
 export enum TkgEventType {
     // vSphere events
-    VC_AUTHENTICATED,
+    VSPHERE_VC_AUTHENTICATED,
     DATACENTER_RESET,
-    DATACENTER_CHANGED,
+    VSPHERE_DATACENTER_CHANGED,
     GET_RESOURCE_POOLS,
     GET_COMPUTE_RESOURCE,
-    GET_VM_NETWORKS,
+    VSPHERE_GET_VM_NETWORKS,
     GET_DATA_STORES,
     GET_VM_FOLDERS,
     VSPHERE_GET_OS_IMAGES,
