+++ pkg/v1/tkg/web/src/app/shared/service/vsphere-wizard-form.service.ts
@@ -10,7 +10,7 @@ import { WizardFormBase } from './wizard-form-base';
 const DataSources = [
     TkgEventType.GET_RESOURCE_POOLS,
     TkgEventType.GET_COMPUTE_RESOURCE,
-    TkgEventType.GET_VM_NETWORKS,
+    TkgEventType.VSPHERE_GET_VM_NETWORKS,
     TkgEventType.GET_DATA_STORES,
     TkgEventType.GET_VM_FOLDERS,
     TkgEventType.VSPHERE_GET_OS_IMAGES
@@ -21,7 +21,7 @@ const DataSpec = {
     // vSphere events
     [TkgEventType.GET_RESOURCE_POOLS]: "getVSphereResourcePools",
     [TkgEventType.GET_COMPUTE_RESOURCE]: "getVSphereComputeResources",
-    [TkgEventType.GET_VM_NETWORKS]: "getVSphereNetworks",
+    [TkgEventType.VSPHERE_GET_VM_NETWORKS]: "getVSphereNetworks",
     [TkgEventType.GET_DATA_STORES]: "getVSphereDatastores",
     [TkgEventType.GET_VM_FOLDERS]: "getVSphereFolders",
     [TkgEventType.VSPHERE_GET_OS_IMAGES]: "getVSphereOSImages"
@@ -31,7 +31,7 @@ const ErrorSpec = {
     // vSphere events
     [TkgEventType.GET_RESOURCE_POOLS]: "Failed to retrieve list of resource pools from the specified vCenter Server.",
     [TkgEventType.GET_COMPUTE_RESOURCE]: "Failed to retrieve list of compute resources from the specified datacenter.",
-    [TkgEventType.GET_VM_NETWORKS]: "Failed to retrieve list of VM networks from the specified vCenter Server.",
+    [TkgEventType.VSPHERE_GET_VM_NETWORKS]: "Failed to retrieve list of VM networks from the specified vCenter Server.",
     [TkgEventType.GET_DATA_STORES]: "Failed to retrieve list of datastores from the specified vCenter Server.",
     [TkgEventType.GET_VM_FOLDERS]: "Failed to retrieve list of vm folders from the specified vCenter Server.",
     [TkgEventType.VSPHERE_GET_OS_IMAGES]: "Failed to retrieve list of OS images from the specified vCenter Server."
@@ -55,7 +55,7 @@ export class VSphereWizardFormService extends WizardFormBase {
         });
 
         // Messenger handlers
-        Broker.messenger.getSubject(TkgEventType.DATACENTER_CHANGED)
+        Broker.messenger.getSubject(TkgEventType.VSPHERE_DATACENTER_CHANGED)
             .subscribe(event => {
                 this.datacenterMoid = event.payload;
                 DataSources.forEach(source => {
