+++ pkg/v1/tkg/web/src/app/views/landing/wizard/shared/components/steps/network-step/network-step.component.html
@@ -17,12 +17,13 @@
     <h4 i18n="network step title" class="inline" aria-level="2">
         Kubernetes Network Settings
     </h4>
+    <!-- Only Vsphere enables the network name -->
     <span aria-live="assertive" *ngIf="enableNetworkName">
         <button class="btn-refresh reset"
                 role="button"
                 aria-label="refresh network"
                 tabindex="0"
-                (click)="loadVSphereNetworks()">
+                (click)="loadNetworks()">
                 <clr-icon
                     *ngIf="!loadingNetworks"
                     shape="refresh"
